+++ src/main/java/io/split/android/client/service/splits/SplitFetcherV2.java
@@ -1,4 +1,4 @@
-package io.split.android.client.backend.splits;
+package io.split.android.client.service.splits;
 
 import io.split.android.client.dtos.SplitChange;
